Avancement

+++ server/controllers/FiliereController.js
@@ -1,4 +1,6 @@
 const { Filiere } = require("../models")
+const { Module } = require("../models")
+const { Groupe } = require("../models")
 
 
 const createFiliere = async (req, res) => {
@@ -44,10 +46,31 @@ const deleteFiliere = async (req, res) => {
     if (!filiere) {
       return res.status(400).json({ message : "Filière non trouvée !" })
     }
+    const filiereInGroupe = await Groupe.findOne({where : {Filiere_codeFil : codeFil}})
+    const filiereInModule = await Module.findOne({where : {Filiere_codeFil : codeFil}})
+    if (filiereInGroupe && filiereInModule) {
+      return res.status(400).json({ message : "Filière est afféctée au modules et groupes !" })
+    }
+    if(filiereInModule) {
+      return res.status(400).json({ message : "Filière est afféctée au modules !" })
+    }
+    if(filiereInGroupe) {
+      return res.status(400).json({ message : "Filière est afféctée au groupes !" })
+    }
     await Filiere.destroy({ where : {codeFil : codeFil} })
     res.status(200).json({ message : "Filière supprimée avec succès !" })
   } catch (error) {
-    
+    res.status(500).json({ message: 'Une erreur est survenue lors de la Suppresion de Filiere.' });
+  }
+}
+
+
+const deleteAllFilieres = async (req, res) => {
+  try {
+    await Filiere.destroy({where : {}})
+    return res.status(200).json({ message : "Filières supprimées avec succès !" })
+  } catch (error) {
+    res.status(500).json({ message: 'Une erreur est survenue lors de la suppression des Filieres.' + error });
   }
 }
 
@@ -55,5 +78,6 @@ module.exports = {
   createFiliere,
   getFiliere,
   getFilieres,
-  deleteFiliere
+  deleteFiliere,
+  deleteAllFilieres
 }
